crawl4ai/webscraping_example/playwrite_example_requirements.txt
# Crawl4AI==0.4.247
# python-dotenv==1.0.1
# pydantic==2.10.6

pdf2image
markdownify
playwright
pdfminer.six
pytesseract